Buying edge bug fixes

+++ ZanP.OrderBooks/Models/Data/OrderBook.cs
@@ -8,7 +8,7 @@ namespace ZanP.OrderBooks.Models.Data
         public int? Id { get; set; }
         public DateTime Time { get; set; }
         public string Type { get; set; }
-        public double Amount { get; set; }
-        public double Price { get; set; }
+        public decimal Amount { get; set; }
+        public decimal Price { get; set; }
     }
 }
